fix(ci): split ctx_read/tests.rs to clear the LOC gate
Merging upstream main (#840-#848 batch) brought tests.rs back to 2062
lines, over the frozen-file limit of 2000 the loc-gate.sh gate enforces.
Extract three cohesive, self-contained sections into sibling test
modules, following the precedent already set by tests_windowed.rs:

- tests_cold_stub.rs: the persistent cold-stub tests (#955)
- tests_delta_explicit.rs: the delta-explicit resolver tests (#498)
- tests_gh775.rs: the ranged-read-after-full-read regression tests (#775)

The two remaining #811 disk-streaming tests move into the existing
tests_windowed.rs, which already covers that exact feature.

tests.rs drops to 1483 lines, under the 1500 non-frozen limit, so it's
removed from the loc-gate.sh allowlist entirely per the script's own
ratchet rule.
