+++ rust/src/tools/ctx_read/mod.rs
@@ -21,6 +21,12 @@ pub(crate) use mode::ReadMode;
 #[cfg(test)]
 mod tests;
 #[cfg(test)]
+mod tests_cold_stub;
+#[cfg(test)]
+mod tests_delta_explicit;
+#[cfg(test)]
+mod tests_gh775;
+#[cfg(test)]
 mod tests_windowed;
 
 /// Pre-counted read output carrying the output string, resolved mode,
