+++ rust/src/tools/ctx_read/tests.rs
@@ -1207,7 +1207,7 @@ fn cache_hit_stub_is_byte_stable_across_rereads() {
 // ---------------------------------------------------------------------------
 
 /// Prime the cache with a full read of the file already on disk at `p`.
-fn primed_full_cache(p: &str) -> SessionCache {
+pub(super) fn primed_full_cache(p: &str) -> SessionCache {
     let mut cache = SessionCache::new();
     let _ = handle_with_task_resolved(&mut cache, p, "full", CrpMode::Off, None);
     debug_assert!(
@@ -1435,386 +1435,6 @@ fn conversation_scoped_stub_served_when_no_context() {
     );
 }
 
-// ---------------------------------------------------------------------------
-// Persistent cold stub (#955): after a daemon restart / idle clear the live
-// cache is empty, so an unchanged re-read must be served from the persisted
-// index — but only for the SAME known conversation and an unchanged file. The
-// record is forged directly (modelling one that outlived the restart) and the
-// current conversation is injected, so the assertions are host-independent.
-// ---------------------------------------------------------------------------
-
-/// Primes a real full delivery to capture authentic (hash, mtime, line_count,
-/// file_ref), then forges a persisted record under `conv`. Clears the global
-/// index before priming (so the prime isn't short-circuited by a stale record)
-/// and after (to drop the prime's own write-through) — leaving exactly the one
-/// forged record.
-fn seed_cold_record(p: &str, conv: &str) {
-    crate::core::read_stub_index::clear_for_test();
-    let primed = primed_full_cache(p);
-    let entry = primed.get(p).unwrap();
-    let rec = crate::core::read_stub_index::StubRecord::new(
-        crate::core::pathutil::normalize_tool_path(p),
-        entry.hash.clone(),
-        entry.stored_mtime,
-        entry.line_count,
-        primed.get_file_ref_readonly(p).unwrap_or_default(),
-        Some(conv.to_string()),
-    );
-    crate::core::read_stub_index::clear_for_test();
-    crate::core::read_stub_index::record(rec);
-}
-
-#[test]
-#[serial_test::serial(stub_index)]
-fn cold_fallback_serves_stub_for_same_conversation_after_restart() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("warm.rs");
-    let p = path.to_string_lossy().to_string();
-    std::fs::write(&path, "fn main() { let x = 1; }\n").unwrap();
-
-    seed_cold_record(&p, "conv-a");
-    // Empty cache models a fresh daemon: the warm path misses, cold fallback fires.
-    let cold = SessionCache::new();
-    let out = try_stub_hit_readonly_scoped(&cold, &p, Some("conv-a"));
-    crate::core::read_stub_index::clear_for_test();
-    assert!(
-        out.is_some_and(|o| o.content.contains("[unchanged")),
-        "same-conversation re-read after restart must serve the persisted stub"
-    );
-}
-
-#[test]
-#[serial_test::serial(stub_index)]
-fn cold_fallback_withheld_for_other_conversation() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("warm.rs");
-    let p = path.to_string_lossy().to_string();
-    std::fs::write(&path, "fn main() { let x = 1; }\n").unwrap();
-
-    seed_cold_record(&p, "conv-a");
-    let cold = SessionCache::new();
-    let out = try_stub_hit_readonly_scoped(&cold, &p, Some("conv-b"));
-    crate::core::read_stub_index::clear_for_test();
-    assert!(
-        out.is_none(),
-        "a different conversation must get a cold full read, never a persisted stub"
-    );
-}
-
-#[test]
-#[serial_test::serial(stub_index)]
-fn cold_fallback_withheld_without_conversation_context() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("warm.rs");
-    let p = path.to_string_lossy().to_string();
-    std::fs::write(&path, "fn main() { let x = 1; }\n").unwrap();
-
-    seed_cold_record(&p, "conv-a");
-    let cold = SessionCache::new();
-    // Unlike the WARM path, an absent conversation cannot prove the content is in
-    // the new process's context → no cold stub (the stricter gate keeps #954's
-    // cross-chat hazard closed across restarts).
-    let out = try_stub_hit_readonly_scoped(&cold, &p, None);
-    crate::core::read_stub_index::clear_for_test();
-    assert!(
-        out.is_none(),
-        "absent conversation context must NOT serve a cold persisted stub"
-    );
-}
-
-#[test]
-#[serial_test::serial(stub_index)]
-fn cold_fallback_withheld_when_file_changed() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("warm.rs");
-    let p = path.to_string_lossy().to_string();
-    std::fs::write(&path, "fn main() { let x = 1; }\n").unwrap();
-
-    seed_cold_record(&p, "conv-a");
-    // Content changed during downtime → mtime/md5 mismatch → no stub.
-    std::fs::write(&path, "fn main() { let x = 2; let y = 3; }\n").unwrap();
-    let cold = SessionCache::new();
-    let out = try_stub_hit_readonly_scoped(&cold, &p, Some("conv-a"));
-    crate::core::read_stub_index::clear_for_test();
-    assert!(
-        out.is_none(),
-        "a file changed on disk must get a cold full read, never a stale stub"
-    );
-}
-
-#[test]
-fn delta_explicit_changed_file_diverts_full_reread_to_diff() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("changed.rs");
-    let p = path.to_string_lossy().to_string();
-    std::fs::write(&path, "fn main() {}\n").unwrap();
-
-    let mut cache = primed_full_cache(&p);
-
-    // File changes on disk after the first full read.
-    std::thread::sleep(Duration::from_secs(1));
-    std::fs::write(&path, "fn main() { changed(); }\n").unwrap();
-
-    let decision = resolve_explicit_delta_mode(
-        &cache, &p, "full", /*explicit*/ true, /*fresh*/ false, true,
-    );
-    assert_eq!(
-        decision.mode, "diff",
-        "changed full re-read must divert to diff"
-    );
-    let note = decision
-        .note
-        .expect("a diff diversion must carry an advisory note");
-    assert!(
-        note.contains("[delta-explicit]"),
-        "note tag missing: {note}"
-    );
-    assert!(
-        note.contains("fresh=true"),
-        "note must mention the bypass: {note}"
-    );
-
-    // End-to-end: the engine renders the diff against the FULL cached content.
-    let out = handle_with_task_resolved(&mut cache, &p, "diff", CrpMode::Off, None);
-    assert_eq!(out.resolved_mode, "diff");
-    assert!(
-        out.content.contains("[diff]"),
-        "engine must emit a diff: {}",
-        out.content
-    );
-    assert!(
-        out.content.contains("changed()"),
-        "diff must reflect the new on-disk content: {}",
-        out.content
-    );
-}
-
-#[test]
-fn delta_explicit_changed_lines_request_diverts_to_diff() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("lines.rs");
-    let p = path.to_string_lossy().to_string();
-    std::fs::write(&path, "fn a() {}\nfn b() {}\n").unwrap();
-
-    let cache = primed_full_cache(&p);
-
-    std::thread::sleep(Duration::from_secs(1));
-    std::fs::write(&path, "fn a() { x(); }\nfn b() {}\n").unwrap();
-
-    let decision = resolve_explicit_delta_mode(&cache, &p, "lines:1-1", true, false, true);
-    assert_eq!(
-        decision.mode, "diff",
-        "a changed-file lines: re-read must divert to diff, not re-extract a window"
-    );
-    assert!(decision.note.is_some());
-}
-
-#[test]
-fn delta_explicit_diff_base_is_full_cached_content_not_compressed() {
-    // Fix #2 guard: the diff base must be the full source the cache stored, even
-    // when the most recent read of the file was a COMPRESSED view (map). If the
-    // base were the compressed view, the diff would be garbage.
-    let _iso = crate::core::data_dir::isolated_data_dir();
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("big.rs");
-    let p = path.to_string_lossy().to_string();
-    let mut content = String::new();
-    for i in 0..60 {
-        content.push_str(&format!(
-            "pub fn original_fn_{i}(x: i32) -> i32 {{ x + {i} }}\n"
-        ));
-    }
-    std::fs::write(&path, &content).unwrap();
-
-    let mut cache = SessionCache::new();
-    // Cache the full content, then read a compressed (map) view — last_mode=map,
-    // but the entry still stores the full source.
-    let _ = handle_with_task_resolved(&mut cache, &p, "full", CrpMode::Off, None);
-    let _ = handle_with_task_resolved(&mut cache, &p, "map", CrpMode::Off, None);
-
-    // Change exactly one line on disk.
-    std::thread::sleep(Duration::from_secs(1));
-    let changed = content.replace(
-        "pub fn original_fn_7(x: i32) -> i32 { x + 7 }",
-        "pub fn original_fn_7(x: i32) -> i32 { x + 70707 }",
-    );
-    std::fs::write(&path, &changed).unwrap();
-
-    let out = handle_with_task_resolved(&mut cache, &p, "diff", CrpMode::Off, None);
-    assert!(
-        out.content.contains("[diff]"),
-        "expected a diff: {}",
-        out.content
-    );
-    // The marker appears only if the diff compared against the FULL original
-    // source (a compressed map base would never contain this literal).
-    assert!(
-        out.content.contains("70707"),
-        "diff must be computed against full cached source, got: {}",
-        out.content
-    );
-    // And it must be a one-line edit, not a wholesale replacement of a
-    // compressed base against the full file.
-    assert!(
-        out.content.contains("+1/-1") || out.content.contains("-1/+1"),
-        "single-line change should diff as +1/-1: {}",
-        out.content
-    );
-}
-
-#[test]
-fn delta_explicit_unchanged_lines_collapse_to_full_stub() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("same.rs");
-    let p = path.to_string_lossy().to_string();
-    std::fs::write(&path, "fn a() {}\nfn b() {}\n").unwrap();
-
-    let cache = primed_full_cache(&p);
-
-    // No disk change. A lines: re-read of a fully-delivered file re-emits text
-    // the model holds → collapse to the full-mode stub (no diff, no note).
-    let decision = resolve_explicit_delta_mode(&cache, &p, "lines:1-1", true, false, true);
-    assert_eq!(
-        decision.mode, "full",
-        "unchanged lines: of a full file must collapse to the stub"
-    );
-    assert!(
-        decision.note.is_none(),
-        "a silent stub collapse must not carry a note"
-    );
-}
-
-#[test]
-fn delta_explicit_unchanged_full_reread_is_untouched() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("same.rs");
-    let p = path.to_string_lossy().to_string();
-    std::fs::write(&path, "fn a() {}\n").unwrap();
-
-    let cache = primed_full_cache(&p);
-
-    // An unchanged full re-read already hits the downstream `[unchanged]` stub;
-    // the resolver leaves it untouched.
-    let decision = resolve_explicit_delta_mode(&cache, &p, "full", true, false, true);
-    assert_eq!(decision.mode, "full");
-    assert!(decision.note.is_none());
-}
-
-#[test]
-fn delta_explicit_off_preserves_current_behavior() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("changed.rs");
-    let p = path.to_string_lossy().to_string();
-    std::fs::write(&path, "fn main() {}\n").unwrap();
-
-    let cache = primed_full_cache(&p);
-
-    std::thread::sleep(Duration::from_secs(1));
-    std::fs::write(&path, "fn main() { changed(); }\n").unwrap();
-
-    // enabled=false → the mode is never rewritten, no matter the disk state.
-    let decision =
-        resolve_explicit_delta_mode(&cache, &p, "full", true, false, /*enabled*/ false);
-    assert_eq!(
-        decision.mode, "full",
-        "feature OFF must preserve the requested mode"
-    );
-    assert!(decision.note.is_none());
-
-    let lines = resolve_explicit_delta_mode(&cache, &p, "lines:1-1", true, false, false);
-    assert_eq!(
-        lines.mode, "lines:1-1",
-        "feature OFF must not touch lines: either"
-    );
-    assert!(lines.note.is_none());
-}
-
-#[test]
-fn delta_explicit_fresh_bypasses() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("changed.rs");
-    let p = path.to_string_lossy().to_string();
-    std::fs::write(&path, "fn main() {}\n").unwrap();
-
-    let cache = primed_full_cache(&p);
-
-    std::thread::sleep(Duration::from_secs(1));
-    std::fs::write(&path, "fn main() { changed(); }\n").unwrap();
-
-    // fresh=true → always bypass even with the feature on and a changed file.
-    let decision = resolve_explicit_delta_mode(&cache, &p, "full", true, /*fresh*/ true, true);
-    assert_eq!(
-        decision.mode, "full",
-        "fresh=true must bypass the diff diversion"
-    );
-    assert!(decision.note.is_none());
-}
-
-#[test]
-fn delta_explicit_first_read_unaffected() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("new.rs");
-    let p = path.to_string_lossy().to_string();
-    std::fs::write(&path, "fn main() {}\n").unwrap();
-
-    // Nothing cached yet — the very first read can never be a diff.
-    let cache = SessionCache::new();
-    let decision = resolve_explicit_delta_mode(&cache, &p, "full", true, false, true);
-    assert_eq!(
-        decision.mode, "full",
-        "an uncached first read must be served normally"
-    );
-    assert!(decision.note.is_none());
-
-    let lines = resolve_explicit_delta_mode(&cache, &p, "lines:1-1", true, false, true);
-    assert_eq!(lines.mode, "lines:1-1");
-    assert!(lines.note.is_none());
-}
-
-#[test]
-fn delta_explicit_only_fires_for_explicit_mode() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("changed.rs");
-    let p = path.to_string_lossy().to_string();
-    std::fs::write(&path, "fn main() {}\n").unwrap();
-
-    let cache = primed_full_cache(&p);
-
-    std::thread::sleep(Duration::from_secs(1));
-    std::fs::write(&path, "fn main() { changed(); }\n").unwrap();
-
-    // explicit_mode=false (mode was auto-resolved) → never diverted; auto-mode
-    // already has its own staleness handling.
-    let decision =
-        resolve_explicit_delta_mode(&cache, &p, "full", /*explicit*/ false, false, true);
-    assert_eq!(
-        decision.mode, "full",
-        "auto-resolved modes must not be diverted to diff"
-    );
-    assert!(decision.note.is_none());
-}
-
-#[test]
-fn delta_explicit_decision_is_byte_stable() {
-    // #498 determinism: the resolver's note carries no timestamp/counter, so
-    // repeated calls on the same changed-file state are byte-identical.
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("changed.rs");
-    let p = path.to_string_lossy().to_string();
-    std::fs::write(&path, "fn main() {}\n").unwrap();
-
-    let cache = primed_full_cache(&p);
-    std::thread::sleep(Duration::from_secs(1));
-    std::fs::write(&path, "fn main() { changed(); }\n").unwrap();
-
-    let d1 = resolve_explicit_delta_mode(&cache, &p, "full", true, false, true);
-    let d2 = resolve_explicit_delta_mode(&cache, &p, "full", true, false, true);
-    assert_eq!(
-        d1, d2,
-        "delta-explicit decision drifted between identical calls"
-    );
-}
-
 #[test]
 fn compress_protect_glob_forces_full_verbatim_read() {
     // #1150: a path matching a `compress_protect` glob is returned verbatim even
@@ -1860,203 +1480,4 @@ fn compress_protect_glob_forces_full_verbatim_read() {
 
     crate::core::config::Config::update_global(|c| c.proxy.compress_protect = None).unwrap();
     crate::test_env::remove_var("LEAN_CTX_SHOW_SAVINGS");
-}
-
-// -- Regression: GitHub Issue #775 --
-// After a full-file read is cached, a subsequent ranged read with `lines:N-M`
-// must return only the requested window — not the full file content again.
-
-/// Helper: create a test file with `n` numbered lines ("line 1\nline 2\n…").
-fn write_numbered_file(dir: &std::path::Path, name: &str, n: usize) -> String {
-    let path = dir.join(name);
-    let content: String = (1..=n)
-        .map(|i| format!("line {i}"))
-        .collect::<Vec<_>>()
-        .join("\n");
-    std::fs::write(&path, &content).unwrap();
-    path.to_string_lossy().to_string()
-}
-
-#[test]
-fn gh775_full_then_ranged_returns_only_window() {
-    let _iso = crate::core::data_dir::isolated_data_dir();
-    crate::test_env::set_var("LEAN_CTX_SHOW_SAVINGS", "0");
-
-    let dir = tempfile::tempdir().unwrap();
-    let p = write_numbered_file(dir.path(), "big.ts", 2000);
-
-    let mut cache = SessionCache::new();
-
-    // 1. Full read — delivers all 2000 lines, marks as fully delivered.
-    let full = handle_with_task_resolved(&mut cache, &p, "full", CrpMode::Off, None);
-    assert!(
-        full.content.contains("line 1"),
-        "full read must include first line"
-    );
-    assert!(
-        full.content.contains("line 2000"),
-        "full read must include last line"
-    );
-
-    // 2. Ranged read — must return ONLY lines 1480–1489.
-    let ranged =
-        handle_fresh_with_task_resolved(&mut cache, &p, "lines:1480-1489", CrpMode::Off, None);
-    assert!(
-        ranged.content.contains("line 1480"),
-        "ranged read must contain the first requested line:\n{}",
-        &ranged.content[..ranged.content.len().min(300)]
-    );
-    assert!(
-        ranged.content.contains("line 1489"),
-        "ranged read must contain the last requested line"
-    );
-    assert!(
-        !ranged.content.contains("line 1\n") && !ranged.content.contains("line 2000"),
-        "ranged read must NOT contain lines outside the window:\n{}",
-        &ranged.content[..ranged.content.len().min(500)]
-    );
-
-    crate::test_env::remove_var("LEAN_CTX_SHOW_SAVINGS");
-}
-
-#[test]
-fn gh775_full_then_ranged_with_fresh_returns_only_window() {
-    let _iso = crate::core::data_dir::isolated_data_dir();
-    crate::test_env::set_var("LEAN_CTX_SHOW_SAVINGS", "0");
-
-    let dir = tempfile::tempdir().unwrap();
-    let p = write_numbered_file(dir.path(), "big2.ts", 2000);
-
-    let mut cache = SessionCache::new();
-
-    // 1. Full read.
-    handle_with_task_resolved(&mut cache, &p, "full", CrpMode::Off, None);
-
-    // 2. Fresh ranged read (simulates `fresh:true` in the tool args).
-    let ranged =
-        handle_fresh_with_task_resolved(&mut cache, &p, "lines:1480-1489", CrpMode::Off, None);
-    assert!(
-        ranged.content.contains("line 1480"),
-        "fresh ranged read must contain requested start line"
-    );
-    assert!(
-        ranged.content.contains("line 1489"),
-        "fresh ranged read must contain requested end line"
-    );
-    assert!(
-        !ranged.content.contains("line 1\n") && !ranged.content.contains("line 2000"),
-        "fresh ranged read must NOT contain lines outside the window"
-    );
-
-    crate::test_env::remove_var("LEAN_CTX_SHOW_SAVINGS");
-}
-
-#[test]
-fn gh775_ranged_response_starts_at_requested_line() {
-    let _iso = crate::core::data_dir::isolated_data_dir();
-    crate::test_env::set_var("LEAN_CTX_SHOW_SAVINGS", "0");
-
-    let dir = tempfile::tempdir().unwrap();
-    let p = write_numbered_file(dir.path(), "big3.ts", 2000);
-
-    let mut cache = SessionCache::new();
-
-    // Full read to warm cache.
-    handle_with_task_resolved(&mut cache, &p, "full", CrpMode::Off, None);
-
-    // Ranged read.
-    let ranged =
-        handle_fresh_with_task_resolved(&mut cache, &p, "lines:500-504", CrpMode::Off, None);
-
-    // The first numbered output line must be line 500.
-    // extract_line_range formats as " 500| line 500".
-    let body_lines: Vec<&str> = ranged
-        .content
-        .lines()
-        .filter(|l| l.contains("| line "))
-        .collect();
-    assert!(
-        !body_lines.is_empty(),
-        "ranged read must contain numbered output lines"
-    );
-    assert!(
-        body_lines[0].contains("500| line 500"),
-        "first body line must be line 500, got: {}",
-        body_lines[0]
-    );
-    assert_eq!(
-        body_lines.len(),
-        5,
-        "lines:500-504 must return exactly 5 lines, got {}",
-        body_lines.len()
-    );
-
-    crate::test_env::remove_var("LEAN_CTX_SHOW_SAVINGS");
-}
-
-#[test]
-fn gh775_cold_ranged_read_returns_only_window() {
-    let _iso = crate::core::data_dir::isolated_data_dir();
-    crate::test_env::set_var("LEAN_CTX_SHOW_SAVINGS", "0");
-
-    let dir = tempfile::tempdir().unwrap();
-    let p = write_numbered_file(dir.path(), "cold.ts", 2000);
-
-    let mut cache = SessionCache::new();
-
-    // No prior full read — cold ranged read must still return only the window.
-    let ranged = handle_with_task_resolved(&mut cache, &p, "lines:100-109", CrpMode::Off, None);
-    assert!(
-        ranged.content.contains("line 100"),
-        "cold ranged read must contain start line"
-    );
-    assert!(
-        ranged.content.contains("line 109"),
-        "cold ranged read must contain end line"
-    );
-    assert!(
-        !ranged.content.contains("line 2000"),
-        "cold ranged read must NOT contain lines outside window"
-    );
-
-    crate::test_env::remove_var("LEAN_CTX_SHOW_SAVINGS");
-}
-
-// ---------------------------------------------------------------------------
-// #811: the `anchored:N-M` disk-streaming short-circuit. A fresh, explicitly
-// windowed anchored read must never materialize the whole file — these tests
-// exercise `read_line_window`/`parse_disk_anchor_range` directly, and the
-// end-to-end case that motivated it: a file bigger than `LCTX_MAX_READ_BYTES`
-// must still serve a small anchored window instead of erroring "file too
-// large" (the bug that `read_file_lossy`'s own error message told the caller
-// to route around, but couldn't actually satisfy before this fix).
-// ---------------------------------------------------------------------------
-
-#[test]
-fn parse_disk_anchor_range_accepts_dash_form_only() {
-    assert_eq!(parse_disk_anchor_range("5-10"), Some((5, 10)));
-    assert_eq!(parse_disk_anchor_range("1-999999"), Some((1, 999_999)));
-    // Bare "N" (meaning "to EOF") needs a known total to resolve — the
-    // streaming path doesn't have one up front, so it declines rather than
-    // guess, and the caller falls back to the full-read path.
-    assert_eq!(parse_disk_anchor_range("5"), None);
-    assert_eq!(parse_disk_anchor_range("not-a-range"), None);
-}
-
-#[test]
-fn read_line_window_streams_only_the_requested_span() {
-    let dir = tempfile::tempdir().unwrap();
-    let path = dir.path().join("window.txt");
-    let body = (1..=50)
-        .map(|i| format!("line {i}"))
-        .collect::<Vec<_>>()
-        .join("\n");
-    std::fs::write(&path, format!("{body}\n")).unwrap();
-    let p = path.to_string_lossy().to_string();
-
-    let window = read_line_window(&p, 10, 12).expect("streamed read must succeed");
-    assert_eq!(window.total_lines, 50);
-    assert_eq!(window.start, 10);
-    assert_eq!(window.end, 12);
-    assert_eq!(window.body, "line 10\nline 11\nline 12");
 }
